[UPDATE] ProductPost Complete

+++ src/main/java/com/treemarket/tree/mapper/CategoryMapper.java
@@ -4,5 +4,5 @@ import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
 public interface CategoryMapper {
-    int getCtgId (String ctgName);
+    Long getCtgId (String ctgName);
 }
